pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "autogen-agentchat>=0.7.5",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "feedparser>=6.0.12",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "openai>=2.32.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.38",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.46.0",
    "websockets>=16.0",
]
[tool.isort]
known_first_party = ["app"]

[tool.ruff]
src = ["app"]
